+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - smart http client body streaming and addHeader method
 - 'enabled' flag in verticle registry
 - plugins to build and upload BOM files
+- ext config stores can control 'ssl' flag of underlying store with 'scheme' attribute
 
 ### Changed
 - registry logger name contains registry type